backup_repository/fully_recovered_duri_system_2025-09-04_212342/duri_control/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
psutil
psycopg2-binary
docker
APScheduler 
PyJWT 